+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 ### Fixed
 * Double-quote titles with square brackets or colons to decode them correctly.
+* Fix default path for TODOs from current directory to $HOME/.todos.
 
 ## 0.5.0 - 2021-08-07